fix(router): exclude already-tried deployments from retry selection
Retryable failures inside execute_with_selected_deployment /
execute_stream_with_selected_deployment recorded cooldown but did not
steer per-request selection, so PriorityBased setups re-picked the same
failed deployment on every attempt and never failed over to a healthy
candidate.

Deployment exclusion is now split by failure kind:

- budget/unpriced exclusions stay hard: when every candidate is excluded
  the loop fails closed with last_error (previous behavior, covered by
  unpriced_model_fallback tests);
- retryable-failure exclusions are soft: selection prefers deployments
  not yet tried this request and falls back to the pool minus budget
  exclusions once every candidate was tried once, so single-deployment
  setups keep same-target retries.

The oversized inline test module moved to execution_tests.rs via
#[path] (same convention as execution_retry_delay_tests.rs). New tests:
unary/stream failover to an untried deployment under high allowed_fails,
and single-deployment same-target retry preservation.
